fix: 달성률 계산은 float이되, 반환은 정수값이어야 함

+++ athena/src/main/java/goorm/athena/domain/project/repository/query/ProjectQueryRepository.java
@@ -53,7 +53,7 @@ public class ProjectQueryRepository {
                         project.seller.nickname,
                         project.title,
                         project.description,
-                        Expressions.numberTemplate(Long.class,
+                        Expressions.numberTemplate(Integer.class,
                                 "floor(({0} * 100.0) / nullif({1}, 0))",
                                 project.totalAmount, project.goalAmount),
                         project.createdAt,
